move tests functions
